+++ components/common/PageTitle.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const PageTitle = styled.h1`
+  width: 100%;
+  display: flex;
+  font-size: 3em;
+  color: #5b5b5b;
+`;
+
+export default PageTitle;